Optimize compiler flags for arm64 and arm64-vdso
This commit updates the compiler flags for arm64 and arm64-vdso code to improve performance. Specifically, the -O2 flag has been changed to -O3 in the Makefile for arm64 code, and in the Makefile for arm64-vdso32 code. These flags control the level of optimization performed by the compiler, with -O3 indicating a higher level of optimization compared to -O2. The change is expected to result in better performance for the respective code.

The commit also adds additional compiler flags for arm64 code when using the clang compiler. These flags are related to the Polly optimization framework and are aimed at improving loop optimizations and vectorization.

Overall, this commit aims to optimize the compiler flags for arm64 and arm64-vdso code, leading to improved performance.